pyproject.toml
[tool.poetry]
name = "shiopa-backend"
version = "0.1.0"
description = ""
authors = ["sheyziSilver <blogsedap@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = "^0.15.0"
fastapi = "^0.68.1"
tortoise-orm = "^0.17.6"
aerich = "^0.5.7"
asyncpg = "^0.24.0"
cloudinary = "^1.26.0"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
